Split core list item components into their own file for reusability
